refactor: move src to server folder
