+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "cross-env NODE_ENV=development concurrently \"next dev\" \"node --import ./scripts/register-ts-node.mjs src/app.ts\"",
+    "dev": "cross-env NODE_ENV=development concurrently \"next dev\" \"node --import ./scripts/register-ts-node.mjs server/src/app.ts\"",
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
@@ -13,21 +13,27 @@
     "build-storybook": "storybook build"
   },
   "dependencies": {
+    "@eslint/config-array": "^0.16.0",
+    "@eslint/object-schema": "^2.1.4",
     "@hookform/resolvers": "^3.4.2",
     "@radix-ui/react-checkbox": "^1.0.4",
     "@radix-ui/react-label": "^2.0.2",
     "@radix-ui/react-radio-group": "^1.1.3",
     "@radix-ui/react-slot": "^1.0.2",
+    "@typescript-eslint/eslint-plugin": "^7.13.0",
+    "@typescript-eslint/parser": "^7.13.0",
     "bcrypt": "^5.1.1",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
+    "eslint-plugin-import": "^2.29.1",
     "express": "^4.19.2",
     "express-rate-limit": "^7.2.0",
-    "express-validator": "^7.0.1",
+    "express-validator": "^7.1.0",
     "firebase-admin": "^12.1.0",
+    "glob": "^10.4.1",
     "helmet": "^7.1.0",
     "jsonwebtoken": "^9.0.2",
     "lucide-react": "^0.379.0",
@@ -69,7 +75,6 @@
     "concurrently": "^8.2.2",
     "cross-env": "^7.0.3",
     "daisyui": "^4.11.1",
-    "eslint": "^8",
     "eslint-config-next": "14.2.3",
     "eslint-plugin-storybook": "^0.8.0",
     "jest": "^29.7.0",
@@ -81,7 +86,9 @@
     "supertest": "^7.0.0",
     "tailwindcss": "^3.4.1",
     "ts-jest": "^29.1.4",
+    "ts-node": "^10.9.2",
     "tsconfig-paths": "^4.2.0",
+    "tsx": "^4.15.2",
     "typescript": "^5"
   },
   "type": "module"
