+++ server/src/interfaces/IUser.ts
