+++ server/src/middlewares/authenticateToken.ts
