+++ server/src/middlewares/pointsMiddleware.ts
