+++ server/src/models/Medication.ts
