+++ server/src/routers/authRoutes.ts
