+++ server/src/routers/medicationRoutes.ts
